Different prevention of duplicate test PyPI uploads
If not triggered by a tag-create event, delete all tags pointing to the
current commit so that it is always treated as a development version.

This prevents a race condition where a commit tagged soon after being
pushed may or may either end on on test PyPI as a .dev version *and* a
release, or only the latter, depending on timing. If there is a matrix
of jobs running, which of these happens could be different for each job,
leading to some platforms having an extra dev version on test PyPI that
others platforms don't.

This makes it deterministic: unless we're in a tag event, don't use any
tags from the current commit, and determine the version from previous
tags only.

And we can get rid of the if statement that was designed to prevent
duplicate PyPI uploads.