Build impure Python packages and conda packages
Add env vars PURE and NOARCH to specify if the package is a pure Python
package, and if it has platform/python-version-independent dependencies
respectively

Extend the build job so that it:
- Builds wheels on all platforms except Linux if PURE == false
- Builds conda packages on all platforms if NOARCH == false

Add manylinux job that builds impure Linux wheels on the PyPA-provided
docker image

Add upload steps for Anaconda cloud

branch name of the workflow is currently set to this branch, and uploads
to test PyPI and Anaconda test channel are commented out for testing.
The repo does not have an anaconda api token associated with it yet in
any case.

Addresses issue #17