Run conda build via cmd.exe on Windows.
It turns out (almost) all our conda woes are caused by using it in git
bash on Windows. Running it in `cmd.exe` instead, the `--from-wheel`
workaround is not necessary.

Regardless of the `--from-wheel` workaround, `setuptools-conda` has
grown the ability to install build dependencies from a `pyproject.toml`
or `setup.cfg`, and I was not able to get it to run `conda install` in a
subprocess without mysterious and subtle failures.

This fix uses `setuptools-conda`'s new invocation as
```bash
$ setuptools-conda build [args] .
```

And has a separate step for Windows and Unix so that cmd.exe can be used
in Windows.