Testing merge of master into maintenance (#26)
* Build impure Python packages and conda packages

Add env vars PURE and NOARCH to specify if the package is a pure Python
package, and if it has platform/python-version-independent dependencies
respectively

Extend the build job so that it:
- Builds wheels on all platforms except Linux if PURE == false
- Builds conda packages on all platforms if NOARCH == false

Add manylinux job that builds impure Linux wheels on the PyPA-provided
docker image

Add upload steps for Anaconda cloud

branch name of the workflow is currently set to this branch, and uploads
to test PyPI and Anaconda test channel are commented out for testing.
The repo does not have an anaconda api token associated with it yet in
any case.

Addresses issue #17

* Switch branch back to master and uncomment test repo uploads.

This will not work unless the rpanderson anaconda cloud account is
created and an API key associated with the repo.

* Minor formatting

* Run conda build via cmd.exe on Windows.

It turns out (almost) all our conda woes are caused by using it in git
bash on Windows. Running it in `cmd.exe` instead, the `--from-wheel`
workaround is not necessary.

Regardless of the `--from-wheel` workaround, `setuptools-conda` has
grown the ability to install build dependencies from a `pyproject.toml`
or `setup.cfg`, and I was not able to get it to run `conda install` in a
subprocess without mysterious and subtle failures.

This fix uses `setuptools-conda`'s new invocation as
```bash
$ setuptools-conda build [args] .
```

And has a separate step for Windows and Unix so that cmd.exe can be used
in Windows.

* Set branch and uploads back to normal

* Install build requirements and build using pep517

PEP517 requires projects to list their build dependencies and build
system (setuptools for us) in pyproject.toml. Build tools like pip can
then know how to build the package, installing their build dependencies
automatically (in an isolated environment such that they are not
installed in your current environment) and then running the build system
to build a wheel or sdist.

Most people use pip to build wheels for projects that use this
mechanism. But pip hasn't implemented the sdist part of it yet, which
would leave us still manually installing dependencies just to make
sdists.

So whilst the Python packaging authority makes up its mind where the
general-purpose PEP517 build commands belong, we use the reference
implementation package `pep517` (which IMHO should just be given a
better name and become the blessed tool for PEP517 builds).

* Restore assurance that we have the latest pip, setuptools and wheel

* Do not need pre-release setuptools_scm now that 4.1 is out.

* Use fetch-depth: 0 in checkout

The checkout action now fetches all tags if you do this.

Remove our own unshallow and rename the step it was in to "Ignore tags"
since that's what it does now.

* Re-enable upload and normal branches

* Re-enable upload and normal branches

Co-authored-by: chrisjbillington <chrisjbillington@gmail.com>
Co-authored-by: Russell Anderson <5637107+rpanderson@users.noreply.github.com>