Correct issues leading to failed mac and Linux wheels
Firstly, running commands with the shell `bash -l {0}`, which is
required for miniconda to be activated, results in the `python` command
being Python 2 on macos. This means on macos the workflow has been
building Python 2 wheels exclusively 😬.

Secondly, the latest version of the `manylinux` action saves wheels to
`dist/` rather than `wheelhouse/`. So we were not even uploading them.

These issues broke macos and Linux wheels for any non-pure packages
(pure packages only a single wheel is needed as build on ubuntu,
without the manylinux action).