Trusted Publisher uploads to PyPI and TestPyPI
Since we can no longer use the presence of TestPyPI and PyPI API keys to
know if we should attempt to upload releases (i.e. we're not running in
a fork), we instead do need some explicit configuration.

Rather than do this all in repository variables, we do it in a config
file in the workflow directory.

RELEASE_REPO is set to the repository releases should be made from
(otherwise uploads are skipped).

For consistency, ANACONDA_USER is set here too rather than as a
repository variable.

For completeness everything else is configurable there too.

Re-add the "ignore tags" step - with fixes since it was incorrect given
recent changes to the workflow `on:` configuration. It is needed due to
race conditions - we are cloning the repo in multiple jobs and a tag
could appear in between them, leading to version inconsistencies.