Enable verbose output for PyPI uploads and skipping existing distribution files
Particularly with TestPyPI uploads, the version number can change in
unexpected ways when tags are added after the fact, allowing for
collisions. Just skip in this case.

For PyPI uploads we don't expect collisions, but still allow skipping
instead of erroring in case e.g. some uploads fail and we need to retry
a workflow job